Move BucketHelper into the util package
